Target System + Click Move 50% working?
Basically because there were 2 separate if statements it wouldn't set obj_player.targeting correctly?

I reversed them and then it would work for the mob but not the friendly.

So I combined them and now it works fine. But when you click friendly to mob, the friendly arrow will go there too.